requirements-gpu.txt

torch>=2.0.0
onnx
onnxruntime-gpu
onnxconverter-common
transformers
huggingface_hub
